HT-2907: move loading classes into Loader module pt.2
