+++ lib/loader/file_loader.rb
