+++ lib/loader/holding_loader.rb
