+++ lib/loader/ht_item_loader.rb
