+++ lib/loader/loaded_file.rb
