+++ lib/loader/member_holding.rb
