+++ lib/loader/member_holding_header.rb
